gcc warnings

+++ src/strlcat.c
@@ -45,7 +45,7 @@ local_strlcat(char *dst, const char *src, size_t siz)
 	/* Find the end of dst and adjust bytes left but don't go past end */
 	while (n-- != 0 && *d != '\0')
 		d++;
-	dlen = d - dst;
+	dlen = (size_t)(d - dst);
 	n = siz - dlen;
 
 	if (n == 0)
@@ -59,5 +59,5 @@ local_strlcat(char *dst, const char *src, size_t siz)
 	}
 	*d = '\0';
 
-	return(dlen + (s - src));	/* count does not include NUL */
+	return ((size_t)dlen + (size_t)(s - src));	/* count does not include NUL */
 }
